fix: gateway 서버 모듈에서 security 제거
