+++ gateway-server/build.gradle.kts
@@ -11,11 +11,9 @@ ext {
 }
 
 dependencies {
-    implementation(project(":domain-core"))
     implementation(project(":common"))
     implementation ("org.springframework.cloud:spring-cloud-starter-gateway")
     implementation ("org.springframework.cloud:spring-cloud-starter-netflix-eureka-client")
-    implementation("org.springframework.boot:spring-boot-starter-security")
     implementation("com.auth0:java-jwt:4.4.0")
 }
