+++ gateway-server/src/main/kotlin/com/xircle/gatewayserver/GatewayServerApplication.kt
@@ -7,7 +7,6 @@ import org.springframework.context.annotation.ComponentScan
 @ComponentScan(
     basePackages = [
         "com.xircle.gatewayserver",
-        "com.xircle.core",
         "com.xircle.common"
     ]
 )
