--- gateway-server/src/main/kotlin/com/xircle/gatewayserver/security/MemberDetails.kt
@@ -1,28 +1,0 @@
-package com.xircle.gatewayserver.security
-
-import com.xircle.core.domain.member.model.Member
-import org.springframework.security.core.GrantedAuthority
-import org.springframework.security.core.authority.SimpleGrantedAuthority
-import org.springframework.security.core.userdetails.UserDetails
-
-class MemberDetails(
-    private val member: Member
-) : UserDetails {
-    override fun getAuthorities(): MutableCollection<out GrantedAuthority> {
-        val authorities = ArrayList<GrantedAuthority>()
-        authorities.add(SimpleGrantedAuthority(member.role.name))
-        return authorities
-    }
-
-    override fun getPassword(): String {
-        return member.password
-    }
-
-    override fun getUsername(): String {
-        return member.email
-    }
-
-    fun getId(): Long? {
-        return member.id
-    }
-}
