+++ gateway-server/src/main/resources/application.yml
@@ -20,3 +20,5 @@ spring:
 server:
   port: 8765
 
+jwt:
+  secret: xircle
